Extract swiftui wrapper to dedicated file.
